problem with code generation solved

+++ db/common.sql
@@ -1,6 +1,6 @@
 CREATE SEQUENCE "language_id";
 CREATE TABLE "language" (
-    "id" BIGINT NOT NULL default nextval('language_id'),
+    "id" INTEGER NOT NULL default nextval('language_id'),
     "name" CHARACTER VARYING(16) NOT NULL,
     "code" CHARACTER VARYING(2) NOT NULL,
     "native" CHARACTER VARYING(16) NOT NULL,
@@ -13,7 +13,7 @@ insert into "language" (name) values ('Russian', 'ru', 'Русский', true);
 
 CREATE SEQUENCE "unit_id";
 CREATE TABLE "unit" (
-    "id" BIGINT NOT NULL default nextval('unit_id'),
+    "id" INTEGER NOT NULL default nextval('unit_id'),
     "name" CHARACTER VARYING(16) NULL,
     "sign" CHARACTER VARYING(16) NULL,
     PRIMARY KEY("id")
@@ -21,8 +21,8 @@ CREATE TABLE "unit" (
 CREATE SEQUENCE "unit_i18n_id";
 CREATE TABLE "unit_i18n" (
     "id" BIGINT NOT NULL default nextval('unit_i18n_id'),
-    "object_id" BIGINT NOT NULL REFERENCES "unit"("id") ON UPDATE CASCADE ON DELETE CASCADE,
-    "language_id" BIGINT NOT NULL REFERENCES "language"("id") ON UPDATE CASCADE ON DELETE CASCADE,
+    "object_id" INTEGER NOT NULL REFERENCES "unit"("id") ON UPDATE CASCADE ON DELETE CASCADE,
+    "language_id" INTEGER NOT NULL REFERENCES "language"("id") ON UPDATE CASCADE ON DELETE CASCADE,
     "name" CHARACTER VARYING(16) NOT NULL,
     PRIMARY KEY("id")
 );
@@ -33,7 +33,7 @@ CREATE UNIQUE INDEX unit_i18n_object_id_language_id_uidx ON "unit_i18n"("object_
 
 CREATE SEQUENCE "city_id";
 CREATE TABLE "city" (
-    "id" BIGINT NOT NULL default nextval('unit_id'),
+    "id" INTEGER NOT NULL default nextval('unit_id'),
     "name" CHARACTER VARYING(16) NULL,
     "latitude" numeric(10,6) NULL,
     "longitude" numeric(10,6) NULL,
@@ -43,8 +43,8 @@ CREATE TABLE "city" (
 CREATE SEQUENCE "city_i18n_id";
 CREATE TABLE "city_i18n" (
     "id" BIGINT NOT NULL default nextval('city_i18n_id'),
-    "object_id" BIGINT NOT NULL REFERENCES "city"("id") ON UPDATE CASCADE ON DELETE CASCADE,
-    "language_id" BIGINT NOT NULL REFERENCES "language"("id") ON UPDATE CASCADE ON DELETE CASCADE,
+    "object_id" INTEGER NOT NULL REFERENCES "city"("id") ON UPDATE CASCADE ON DELETE CASCADE,
+    "language_id" INTEGER NOT NULL REFERENCES "language"("id") ON UPDATE CASCADE ON DELETE CASCADE,
     "name" CHARACTER VARYING(16) NULL,
     PRIMARY KEY("id")
 );
